+++ CHANGES.md
@@ -1,5 +1,5 @@
-NEXT
-----
+1.0.9
+-----
 - drf : improve missing introspection endpoint error handling
 - have all `django_pyoidc` exceptions inherit from `DjangoOIDCException`
 - improve documentation for drf integration : `use_introspection_on_access_tokens` and `hook_validate_access_token`